revolucion/requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cleanup==4.0.0
django-debug-toolbar==2.2
django-filter==2.2.0
djangorestframework==3.11.0
Faker==0.9.1
idna==2.8
Markdown==3.2.2
mixer==6.1.3
pprint==0.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.2
urllib3==1.25.7
